+++ app/helpers/mobile_helper.rb
@@ -7,25 +7,30 @@ module MobileHelper
         reshare = Reshare.where(author_id: current_user.person_id,
                                 root_guid: absolute_root.guid).first
         klass = reshare.present? ? "active" : "inactive"
-        link_to "", reshares_path(root_guid: absolute_root.guid),
+        link_to content_tag(:span, post.reshares.size, class: "count reshare-count"),
+                reshares_path(root_guid: absolute_root.guid),
                 title: t("reshares.reshare.reshare_confirmation", author: absolute_root.author_name),
                 class: "entypo-reshare reshare-action #{klass}"
       else
-        content_tag :div, nil, class: "entypo-reshare reshare-action disabled"
+        content_tag :div,
+                    content_tag(:span, post.reshares.size, class: "count reshare-count"),
+                    class: "entypo-reshare reshare-action disabled"
       end
     else
-      content_tag :div, nil, class: "entypo-reshare reshare-action disabled"
+      content_tag :div,
+                  content_tag(:span, post.reshares.size, class: "count reshare-count"),
+                  class: "entypo-reshare reshare-action disabled"
     end
   end
 
   def mobile_like_icon(post)
     if current_user && current_user.liked?(post)
-      link_to "",
+      link_to content_tag(:span, post.likes.size, class: "count like-count"),
               "#",
               data:  {url: post_like_path(post.id, current_user.like_for(post).id)},
               class: "entypo-heart like-action active"
     else
-      link_to "",
+      link_to content_tag(:span, post.likes.size, class: "count like-count"),
               "#",
               data:  {url: post_likes_path(post.id)},
               class: "entypo-heart like-action inactive"
@@ -33,7 +38,9 @@ module MobileHelper
   end
 
   def mobile_comment_icon(post)
-    link_to "", new_post_comment_path(post), class: "entypo-comment comment-action inactive"
+    link_to content_tag(:span, post.comments.size, class: "count comment-count"),
+            new_post_comment_path(post),
+            class: "entypo-comment comment-action inactive"
   end
 
   def show_comments_link(post, klass="")